Extract retrieval of integration steps for testability.
