+++ cucumber.eclipse.editor/src/main/java/cucumber/eclipse/editor/editors/FeatureFileUtil.java
@@ -1,14 +1,5 @@
 package cucumber.eclipse.editor.editors;
 
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-
-import org.eclipse.core.resources.IFile;
-import org.eclipse.core.runtime.CoreException;
-import org.eclipse.core.runtime.IConfigurationElement;
-import org.eclipse.core.runtime.Platform;
 import org.eclipse.jface.text.BadLocationException;
 import org.eclipse.jface.text.IDocument;
 import org.eclipse.jface.text.IRegion;
@@ -16,11 +7,7 @@ import org.eclipse.ui.IEditorPart;
 import org.eclipse.ui.texteditor.IDocumentProvider;
 import org.eclipse.ui.texteditor.ITextEditor;
 
-import cucumber.eclipse.steps.integration.IStepDefinitions;
-import cucumber.eclipse.steps.integration.Step;
-
-class FeatureFileUtil{
-	final static String EXTENSION_POINT_STEPDEFINITIONS_ID = "cucumber.eclipse.steps.integration";
+public class FeatureFileUtil{
 
 	static String getDocumentLanguage(IEditorPart editorPart) {
 		String lang = null;
@@ -44,30 +31,5 @@ class FeatureFileUtil{
 		}
 
 		return lang;
-	}
-	
-	static Set<Step> getStepsInEncompassingProject(IFile featurefile) {
-		Set<Step> steps = new HashSet<Step>();
-		for (IStepDefinitions stepDef : FeatureFileUtil.getStepDefinitions()) {
-			steps.addAll(stepDef.getSteps(featurefile));
-		}
-		return steps;
-	}
-
-	private static List<IStepDefinitions> getStepDefinitions() {
-		List<IStepDefinitions> stepDefs = new ArrayList<IStepDefinitions>();
-		IConfigurationElement[] config = Platform
-				.getExtensionRegistry()
-				.getConfigurationElementsFor(FeatureFileUtil.EXTENSION_POINT_STEPDEFINITIONS_ID);
-		try {
-			for (IConfigurationElement ce : config) {
-				Object obj = ce.createExecutableExtension("class");
-				if (obj instanceof IStepDefinitions) {
-					stepDefs.add((IStepDefinitions) obj);
-				}
-			}
-		} catch (CoreException e) {
-		}
-		return stepDefs;
 	}
 }
