+++ cucumber.eclipse.editor/src/main/java/cucumber/eclipse/editor/steps/ExtensionRegistryStepProvider.java
@@ -1,0 +1,44 @@
+package cucumber.eclipse.editor.steps;
+
+import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
+
+import org.eclipse.core.resources.IFile;
+import org.eclipse.core.runtime.CoreException;
+import org.eclipse.core.runtime.IConfigurationElement;
+import org.eclipse.core.runtime.Platform;
+
+import cucumber.eclipse.steps.integration.IStepDefinitions;
+import cucumber.eclipse.steps.integration.Step;
+
+public class ExtensionRegistryStepProvider implements IStepProvider {
+
+	final static String EXTENSION_POINT_STEPDEFINITIONS_ID = "cucumber.eclipse.steps.integration";
+
+	public Set<Step> getStepsInEncompassingProject(IFile featurefile) {
+		Set<Step> steps = new HashSet<Step>();
+		for (IStepDefinitions stepDef : getStepDefinitions()) {
+			steps.addAll(stepDef.getSteps(featurefile));
+		}
+		return steps;
+	}
+
+	private static List<IStepDefinitions> getStepDefinitions() {
+		List<IStepDefinitions> stepDefs = new ArrayList<IStepDefinitions>();
+		IConfigurationElement[] config = Platform
+				.getExtensionRegistry()
+				.getConfigurationElementsFor(EXTENSION_POINT_STEPDEFINITIONS_ID);
+		try {
+			for (IConfigurationElement ce : config) {
+				Object obj = ce.createExecutableExtension("class");
+				if (obj instanceof IStepDefinitions) {
+					stepDefs.add((IStepDefinitions) obj);
+				}
+			}
+		} catch (CoreException e) {
+		}
+		return stepDefs;
+	}
+}
