+++ cucumber.eclipse.editor/src/main/java/cucumber/eclipse/editor/steps/IStepProvider.java
@@ -1,0 +1,13 @@
+package cucumber.eclipse.editor.steps;
+
+import java.util.Set;
+
+import org.eclipse.core.resources.IFile;
+
+import cucumber.eclipse.steps.integration.Step;
+
+public interface IStepProvider {
+
+	Set<Step> getStepsInEncompassingProject(IFile featurefile);
+
+}
